feat(custom): 支付回调处理

+++ live-api/src/main/java/top/flobby/live/api/dto/ProductReqDTO.java
@@ -35,5 +35,5 @@ public class ProductReqDTO {
      *
      * @see top.flobby.live.bank.constant.PayChannelEnum
      */
-    private Byte payChannel;
+    private Integer payChannel;
 }
