+++ live-bank-provider/src/main/java/top/flobby/live/bank/provider/dao/po/PayOrderPO.java
@@ -38,7 +38,7 @@ public class PayOrderPO {
     private Integer productId;
     private Byte status;
     private Long userId;
-    private Byte payChannel;
+    private Integer payChannel;
     /**
      * 来源
      *
